+++ services/device-registry-base/src/main/java/org/eclipse/hono/deviceregistry/util/DeviceRegistryUtils.java
@@ -18,6 +18,7 @@ import java.security.cert.CertificateException;
 import java.security.cert.CertificateFactory;
 import java.security.cert.X509Certificate;
 import java.time.Instant;
+import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
@@ -241,5 +242,71 @@ public final class DeviceRegistryUtils {
                 passwordSecret.verifyHashAlgorithm(hashAlgorithmsWhitelist, maxBcryptIterations);
             }
         }
+    }
+
+    /**
+     * Verifies that the elements of a list of credentials are unique regarding type and authentication
+     * identifier.
+     *
+     * @param credentials The credentials to check.
+     * @return A future indicating the outcome of the check.
+     *         The future will be failed with a {@link ClientErrorException} if the elements are
+     *         not unique.
+     */
+    public static Future<?> assertTypeAndAuthIdUniqueness(final List<? extends CommonCredential> credentials) {
+
+        final long uniqueAuthIdAndTypeCount = credentials.stream()
+            .map(credential -> String.format("%s::%s", credential.getType(), credential.getAuthId()))
+            .distinct()
+            .count();
+
+        if (credentials.size() > uniqueAuthIdAndTypeCount) {
+            return Future.failedFuture(new ClientErrorException(HttpURLConnection.HTTP_BAD_REQUEST,
+                    "credentials must have unique (type, auth-id)"));
+        } else {
+            return Future.succeededFuture();
+        }
+    }
+
+    /**
+     * Checks if the properties in the given client context matches with the extension properties of the
+     * given credential.
+     * <p>
+     * The matching properties in the client context and credential extensions are filtered out based 
+     * on their property names. If their values are equal then {@code true} is returned. Otherwise {@code false}.
+     * <p>
+     * The properties that exist only in the client context but not in the extension properties are ignored.
+     * If none of the properties in the client context are found in the credential extensions,
+     * then {@code true} is returned.
+     *
+     * @param credential The credential object to match.
+     * @param clientContext The client context, which contains properties that can be used to identify the device.
+     * @return {@code true} if the properties from client context matches with those of the given credential,
+     *         {@code false} otherwise.
+     * @throws NullPointerException if credential is {@code null}.
+     */
+    public static boolean matchesWithClientContext(final JsonObject credential,
+            final JsonObject clientContext) {
+
+        Objects.requireNonNull(credential);
+
+        final JsonObject extensionProperties = credential.getJsonObject(RegistryManagementConstants.FIELD_EXT,
+                new JsonObject());
+
+        if (Objects.isNull(clientContext) || clientContext.isEmpty() || extensionProperties.isEmpty()) {
+            return true;
+        }
+
+        return clientContext.stream()
+                .filter(entry -> Objects.nonNull(entry.getValue()))
+                .allMatch(entry -> Optional
+                        .ofNullable(extensionProperties.getValue(entry.getKey()))
+                        .map(value -> {
+                            LOG.debug("comparing client context property [name: {}, value: {}] to value on record: {}",
+                                    entry.getKey(), entry.getValue(), value);
+                            return value.equals(entry.getValue());
+                        })
+                        .orElse(true));
+
     }
 }